Add files via upload
**Modeling of Contact in Biphasic Soft Material**
The purpose of the present study is to develop a contact model of biphasic soft material. The ever-widening demand for soft swollen materials in biomedical and engineering applications contributes toward the inspiration of the computational modeling of their mechanical response. The mechanical properties of biphasic soft swollen materials are dependent on the behavior of two phases: the solid matrix and the interstitial fluid. Biphasic soft materials undergo large deformation and rate-dependent nonlinear mechanical response. The investigation of mechanical responses of biphasic soft material under contact scenarios is essential for reliable performance at the time of use. In particular, the contact evaluation of biphasic material is indispensable to analyze the contact interaction of soft robots during locomotion and gripping, the interaction of flexible soft electronics with human skin and fabric, and the interaction between synthetic transplant with bodily tissues during implantation. 
In this current work, modeling of biphasic-biphasic contact and biphasic-solid contact is conducted using finite element software FEBio to determine the contact properties as well as rate-dependent deformation of biphasic soft material. The contact modeling of biphasic material is the initial step that will lead to the development of enhanced formulations for modeling adhesion and friction response in biological and synthetic multiphasic poroviscoelastic materials. In addition, this contact model can be utilized for multi contact modeling such as musculoskeletal joints, distressing impacts in soft tissues, and interaction between cracked surfaces in self-healing polymer gel and hard gels. For biphasic-biphasic contact analysis, both substrate and indenter are constituted of biphasic material and indentation loading is applied to the substrate through the biphasic indenter. Likewise, in case of biphasic-solid contact, solid indenter is used to apply indentation load on biphasic substrate. Indentation is a commonly used technique to identify the dependency of contact phenomenon of biphasic material on loading rate and the constitutive material of indenter. The length scale parameters (radius of the indenter and depth of the indentation), time scale parameters (diffusion time, relaxation time and process time), loading rate and type of the indenter are altered to observe their effect on the mechanical behavior of the biphasic materials. This study for biphasic-biphasic indentation exhibits a significant difference in local stresses between contact regions to non-contact regions according to St. Venant’s principle. The matrix stress, contact pressure and fluid pressure variation for biphasic-solid contact will be represented as well via this analysis. This poster presentation will describe the above research on combined mechanical characteristics of solid skeleton and interstitial fluid within biphasic material.