Add files via upload
The demand for poroelastic soft materials in biomedical and engineering applications is increasing exponentially due to their remarkable resemblance to biological soft tissues and unique mechanical properties. The study of mechanical responses of poroelastic soft material under contact scenarios is one of the active areas of research which is essential for reliable performance of the parts made with these materials. The characteristics of poroelastic multiphasic soft swollen materials rely on the behavior of the solid matrix and the interstitial fluid. Soft multiphasic materials undergo large deformations and exhibit mechanical response that is both dependent on the rate of deformation and characterized by nonlinearity. However, the existing computational models often overlook the influence of adhesion in the contact formulation, neglecting its significance in representing real-world scenarios. The purpose of this study is to bridge this gap by incorporating the rate-dependent adhesion effects for poroelastic material contact into an open-source nonlinear finite element analysis code FEBio. 
Indentation is a commonly used technique to identify the dependency of contact phenomenon on a broad range of loading and unloading rate, relaxation time, and coupling of adhesion and friction. The main obstacle of contact formulations is to precisely determine the active contact zone within the interface by solving a constrained optimization problem with an appropriate set of Karush-Kuhn-Tucker (KKT) conditions. This research aims to explore the mechanics of rate-dependent adhesion and the associated suction effect within a poroelastic contact model. Our primary objective is the modification of the FEBio source code to incorporate effective traction-based contact formulation with nuanced intricacies of adhesion at interface. In addition, the comprehensive capture of rate-dependent adhesion characteristics is focused on the identification of the active contact zone through the KKT condition. The simulation results following the compilation of the modified source code are validated against experimental results. The adaptation of the FEBio source code to include a rate-dependent adhesion, coupled with the precise determination of the active contact zone through the KKT condition, not only sets out to enrich our understanding of material behavior but also improves the capabilities of computational tools towards more authentic simulations of intricate biological interactions.